+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.2.1",
     "antd": "^5.4.7",
     "daisyui": "^2.51.6",
+    "dotenv": "^16.0.3",
     "firebase": "^9.21.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -43,8 +44,6 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.14",
-    "postcss-loader": "^7.3.0",
     "tailwindcss": "^3.3.2"
   }
 }
