--- src/components/cards.js
@@ -1,34 +1,0 @@
-import React from 'react'
-
-
-export function cards() {
-  return<div className='card' >
-          <div class="card w-96 bg-base-100 shadow-xl">
-    <figure class="px-10 pt-10">
-      <img src={require('../images/her78.PNG')} alt="Shoes" class="rounded-xl" />
-    </figure>
-    <div class="card-body items-center text-center">
-      <h2 class="card-title">Drugs</h2>
-      <p>If a dog chews shoes whose shoes does he choose?</p>
-      <div class="card-actions">
-      <button class="badge badge-outline">Electrician</button>
-      <button class="badge badge-outline">3 years job experience</button>
-      <a> Year started: 1/23/2019  </a> <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
-                            <path strokeLinecap="round" strokeLinejoin="round" d="M6.75 3v2.25M17.25 3v2.25M3 18.75V7.5a2.25 2.25 0 012.25-2.25h13.5A2.25 2.25 0 0121 7.5v11.25m-18 0A2.25 2.25 0 005.25 21h13.5A2.25 2.25 0 0021 18.75m-18 0v-7.5A2.25 2.25 0 015.25 9h13.5A2.25 2.25 0 0121 11.25v7.5" />
-                            </svg>
-      {/* <div className="rating-md ml-46 mt-4"><a>Ratings: </a>
-        <input type="radio" name="rating-2" className="mask mask-star-2 bg-orange-400" />
-        <input type="radio" name="rating-2" className="mask mask-star-2 bg-orange-400" />
-        <input type="radio" name="rating-2" className="mask mask-star-2 bg-orange-400" />
-        <input type="radio" name="rating-2" className="mask mask-star-2 bg-orange-400" />
-        <input type="radio" name="rating-2" className="mask mask-star-2 bg-orange-400" />
-    </div> */}
-        <button class="btn btn-primary ml-24 mt-10">Hire Now</button>
-      </div>
-      
-    </div>
-  </div>
-  
-    </div>  
-}
-export default cards
