+++ src/components/drawerleft.js
@@ -1,17 +1,21 @@
 import React from 'react'
-export function drawerleft() {
+
+function DrawerLeft() {
   return (
     <div className="drawer drawer-mobile">
     <div className="drawer-side">
       <label htmlFor="my-drawer-2" className="drawer-overlay"></label> 
       <ul className="menu p-4 w-80 bg-base-100 text-base-content">
-        <li><a href='Home'>Home</a></li>
-        <li><a>Maps</a></li>
+        <li>
+          <a href='Home'>Home</a>
+        </li>
+        <li>
+          <a>Maps</a>
+        </li>
       </ul>
-    
     </div>
   </div>
   )
 }
 
-export default drawerleft
+export default DrawerLeft;
