+++ src/components/index.js
@@ -1,0 +1,9 @@
+import Card from "./Card";
+import Content from "./Content";
+import DrawerLeft from "./DrawerLeft";
+import Footer from "./Footer";
+import Modal from "./Modal";
+import Navbar from "./Navbar";
+import Pagination from "./Pagination";
+
+export { Card, Content, DrawerLeft, Footer, Modal, Navbar, Pagination };
