+++ src/components/pagination.js
@@ -1,14 +1,16 @@
 import React from 'react'
 
-export function pagination() {
-  return <div className='Pages'>
-    <div className="btn-group text-center" style={{ marginTop: '1rem' }}>
-         <button className="btn">1</button>
-            <button className="btn">2</button>
-              <button className="btn">3</button>
-                 <button className="btn">4</button>
-     </div> 
-  </div>
+function Pagination() {
+  return (
+    <div className='Pages'>
+      <div className="btn-group text-center" style={{ marginTop: '1rem' }}>
+          <button className="btn">1</button>
+          <button className="btn">2</button>
+          <button className="btn">3</button>
+          <button className="btn">4</button>
+      </div> 
+    </div>
+  );
 }
 
-export default pagination
+export default Pagination;
