+++ tailwind.config.js
@@ -2,14 +2,14 @@
 module.exports = {
   content: [
     "./src/**/*.{js,jsx,ts,tsx}",
-    
   ],
   theme: {
     extend: {},
   },
-  plugins: [require("@tailwindcss/typography"),require("daisyui"),
-  require('@tailwindcss/forms'),require('@heroicons/react/24/outline')
-
-],
+  plugins: [
+    require("@tailwindcss/typography"),
+    require("daisyui"),
+    require('@tailwindcss/forms'),
+    require('@heroicons/react/24/outline')
+  ],
 }
-
